\name{deployShinyBuilder}
\alias{deployShinyBuilder}
\title{Deploy ShinyBuilder to a specified directory}
\usage{
deployShinyBuilder(dir, update = T, times = c(0, 0, "*", "*", "*"))
}
\arguments{
  \item{dir}{directory to which ShinyBuilder app is
  deployed}

  \item{update}{logical value indicating whether
  ShinyBuilder dashboards should be periodically updated by
  re-running the queries.  True by default.}

  \item{times}{vector specifying update schedule in crontab
  format (i.e., minute (0-59), hour (0-23), day of month
  (1-31), month (1-12), day of week (0-6, Sun = 6)).
  Defaults to daily updates at 12:00 AM.}
}
\description{
Deploy ShinyBuilder app (ui.R, server.R, global.R) to a
specified directory. Useful for copying ShinyBuilder to a
Shiny Server directory from which it will be served. Note
that regardless of its deploy location, the app will use
the ShinyBuilder library location for saving/loading
dashboards, databases, and config files.
}
\examples{
\dontrun{deployShinyBuilder(dir = '/srv/shiny-server/ShinyBuilder')}
}

